+++ main.go
@@ -39,7 +39,7 @@ import (
 )
 
 var (
-	version = "1.4.9"
+	version = "1.4.10"
 	build   = "dev"
 )